+++ src/pl/plajer/buildbattle3/arena/plots/ArenaPlot.java
@@ -49,8 +49,7 @@ import pl.plajerlair.core.services.ReportedException;
  */
 public class ArenaPlot {
 
-  private Location maxPoint;
-  private Location minPoint;
+  private Cuboid cuboid;
   private int points;
   private List<UUID> uuids = new ArrayList<>();
   private Map<Location, Particle> particles = new HashMap<>();
@@ -89,20 +88,12 @@ public class ArenaPlot {
     this.weatherType = weatherType;
   }
 
-  private Location getMaxPoint() {
-    return maxPoint;
+  public Cuboid getCuboid() {
+    return cuboid;
   }
 
-  public void setMaxPoint(Location maxPoint) {
-    this.maxPoint = maxPoint;
-  }
-
-  private Location getMinPoint() {
-    return minPoint;
-  }
-
-  public void setMinPoint(Location minPoint) {
-    this.minPoint = minPoint;
+  public void setCuboid(Cuboid cuboid) {
+    this.cuboid = cuboid;
   }
 
   public List<UUID> getOwners() {
@@ -119,7 +110,6 @@ public class ArenaPlot {
 
   public void fullyResetPlot() {
     try {
-      Cuboid cuboid = new Cuboid(getMinPoint(), getMaxPoint());
       for(Block block : cuboid.blockList()){
         if(block.getType() != Material.AIR) {
           block.setType(Material.AIR);
@@ -144,8 +134,8 @@ public class ArenaPlot {
         }
       }
       getParticles().clear();
-      for (Entity entity : getCenter().getWorld().getEntities()) {
-        if (isInPlotRange(entity.getLocation(), 3)) {
+      for (Entity entity : cuboid.getCenter().getWorld().getEntities()) {
+        if (cuboid.isInWithMarge(entity.getLocation(), 3)) {
           if (JavaPlugin.getPlugin(Main.class).getServer().getPluginManager().isPluginEnabled("Citizens")) {
             if (CitizensAPI.getNPCRegistry().isNPC(entity)) return;
           }
@@ -162,7 +152,6 @@ public class ArenaPlot {
 
   public void resetPlot() {
     try {
-      Cuboid cuboid = new Cuboid(getMinPoint(), getMaxPoint());
       for(Block block : cuboid.blockList()){
         if(block.getType() != Material.AIR) {
           block.setType(Material.AIR);
@@ -179,8 +168,8 @@ public class ArenaPlot {
           setWeatherType(WeatherType.CLEAR);
         }
       }
-      for (Entity entity : getCenter().getWorld().getEntities()) {
-        if (isInPlotRange(entity.getLocation(), 3)) {
+      for (Entity entity : cuboid.getCenter().getWorld().getEntities()) {
+        if (cuboid.isInWithMarge(entity.getLocation(), 3)) {
           if (JavaPlugin.getPlugin(Main.class).getServer().getPluginManager().isPluginEnabled("Citizens")) {
             if (CitizensAPI.getNPCRegistry().isNPC(entity)) continue;
           }
@@ -195,27 +184,6 @@ public class ArenaPlot {
     }
   }
 
-  public Location getCenter() {
-    double x, y, z;
-    if (getMinPoint().getX() > getMaxPoint().getX()) {
-      x = getMaxPoint().getX() + ((getMinPoint().getX() - getMaxPoint().getX()) / 2);
-    } else {
-      x = getMinPoint().getX() + ((getMaxPoint().getX() - getMinPoint().getX()) / 2);
-    }
-    if (getMinPoint().getY() > getMaxPoint().getY()) {
-      y = getMaxPoint().getY() + ((getMinPoint().getY() - getMaxPoint().getY()) / 2);
-    } else {
-      y = getMinPoint().getY() + ((getMaxPoint().getY() - getMinPoint().getY()) / 2);
-    }
-    if (getMinPoint().getZ() > getMaxPoint().getZ()) {
-      z = getMaxPoint().getZ() + ((getMinPoint().getZ() - getMaxPoint().getZ()) / 2);
-    } else {
-      z = getMinPoint().getZ() + ((getMaxPoint().getZ() - getMinPoint().getZ()) / 2);
-    }
-    return new Location(getMinPoint().getWorld(), x, y, z);
-
-  }
-
   public int getPoints() {
     return points;
   }
@@ -224,41 +192,19 @@ public class ArenaPlot {
     this.points = points;
   }
 
-  public boolean isInPlot(Location location) {
-    boolean trueOrNot = false;
-    if (location.getWorld() == getMinPoint().getWorld() && location.getWorld() == getMaxPoint().getWorld()) {
-      if (location.getX() >= getMinPoint().getX() && location.getX() <= getMaxPoint().getX()) {
-        if (location.getY() >= getMinPoint().getY() && location.getY() <= getMaxPoint().getY()) {
-          if (location.getZ() >= getMinPoint().getZ() && location.getZ() <= getMaxPoint().getZ()) {
-            trueOrNot = true;
-          }
-        }
-      }
-      if (location.getX() <= getMinPoint().getX() && location.getX() >= getMaxPoint().getX()) {
-        if (location.getY() <= getMinPoint().getY() && location.getY() >= getMaxPoint().getY()) {
-          if (location.getZ() <= getMinPoint().getZ() && location.getZ() >= getMaxPoint().getZ()) {
-            trueOrNot = true;
-          }
-        }
-      }
-    }
-    return trueOrNot;
-  }
-
   private void changeFloor(Material material) {
     try {
       double y;
-      if (getMinPoint().getY() > getMaxPoint().getY()) {
-        y = getMaxPoint().getY();
+      if (cuboid.getPoint1().getY() > cuboid.getPoint2().getY()) {
+        y = cuboid.getPoint2().getY();
       } else {
-        y = getMinPoint().getY();
+        y = cuboid.getPoint1().getY();
       }
-      Cuboid cuboid = new Cuboid(getMinPoint(), getMaxPoint());
       Location min = cuboid.getPoint1();
       Location max = cuboid.getPoint2();
       for (int x = min.getBlockX(); x <= max.getBlockX(); x = x + 1) {
         for (int z = min.getBlockZ(); z <= max.getBlockZ(); z = z + 1) {
-          Location tmpblock = new Location(getMaxPoint().getWorld(), x, y, z);
+          Location tmpblock = new Location(cuboid.getPoint2().getWorld(), x, y, z);
           tmpblock.getBlock().setType(material);
         }
       }
@@ -272,17 +218,16 @@ public class ArenaPlot {
       if (material == Material.WATER_BUCKET) material = Material.WATER;
       if (material == Material.LAVA_BUCKET) material = Material.LAVA;
       double y;
-      if (getMinPoint().getY() > getMaxPoint().getY()) {
-        y = getMaxPoint().getY();
+      if (cuboid.getPoint1().getY() > cuboid.getPoint2().getY()) {
+        y = cuboid.getPoint2().getY();
       } else {
-        y = getMinPoint().getY();
+        y = cuboid.getPoint1().getY();
       }
-      Cuboid cuboid = new Cuboid(getMinPoint(), getMaxPoint());
       Location min = cuboid.getPoint1();
       Location max = cuboid.getPoint2();
       for (int x = min.getBlockX(); x <= max.getBlockX(); x = x + 1) {
         for (int z = min.getBlockZ(); z <= max.getBlockZ(); z = z + 1) {
-          Location tmpblock = new Location(getMaxPoint().getWorld(), x, y, z);
+          Location tmpblock = new Location(cuboid.getPoint2().getWorld(), x, y, z);
           tmpblock.getBlock().setType(material);
           tmpblock.getBlock().setData(data);
         }
@@ -295,10 +240,10 @@ public class ArenaPlot {
   public Material getFloorMaterial() {
     try {
       Location location;
-      if (getMinPoint().getY() > getMaxPoint().getY()) {
-        location = getMaxPoint().clone();
+      if (cuboid.getPoint1().getY() > cuboid.getPoint2().getY()) {
+        location = cuboid.getPoint2().clone();
       } else {
-        location = getMinPoint().clone();
+        location = cuboid.getPoint1().clone();
       }
       Material material = location.add(0, -1, 0).getBlock().getType();
       if (material == Material.WATER || material == Material.STATIONARY_WATER) return Material.WATER_BUCKET;
@@ -311,49 +256,9 @@ public class ArenaPlot {
     }
   }
 
-  public boolean isInFlyRange(Player player) {
-    boolean trueOrNot = false;
-    Location location = player.getLocation();
-    if (location.getWorld() == getMinPoint().getWorld() && location.getWorld() == getMaxPoint().getWorld()) {
-      if (location.getX() >= getMinPoint().getX() - 5 && location.getX() <= getMaxPoint().getX() + 5) {
-        if (location.getY() >= getMinPoint().getY() - 5 && location.getY() <= getMaxPoint().getY() + 5) {
-          if (location.getZ() >= getMinPoint().getZ() - 5 && location.getZ() <= getMaxPoint().getZ() + 5) {
-            trueOrNot = true;
-          }
-        }
-      }
-      if (location.getX() <= getMinPoint().getX() + 5 && location.getX() >= getMaxPoint().getX() - 5) {
-        if (location.getY() <= getMinPoint().getY() && location.getY() >= getMaxPoint().getY() - 5) {
-          if (location.getZ() <= getMinPoint().getZ() + 5 && location.getZ() >= getMaxPoint().getZ() - 5) {
-            trueOrNot = true;
-          }
-        }
-      }
-    }
-    return trueOrNot;
-  }
-
-  public boolean isInPlotRange(Location location, int added) {
-    if (location.getWorld() == getMinPoint().getWorld() && location.getWorld() == getMaxPoint().getWorld()) {
-      if (location.getX() >= getMinPoint().getX() - added && location.getX() <= getMaxPoint().getX() + added) {
-        if (location.getY() >= getMinPoint().getY() - added && location.getY() <= getMaxPoint().getY() + added) {
-          if (location.getZ() >= getMinPoint().getZ() - added && location.getZ() <= getMaxPoint().getZ() + added) {
-            return true;
-          }
-        }
-      }
-      if (location.getX() <= getMinPoint().getX() + 5 && location.getX() >= getMaxPoint().getX() - 5) {
-        if (location.getY() <= getMinPoint().getY() && location.getY() >= getMaxPoint().getY() - 5) {
-          return location.getZ() <= getMinPoint().getZ() + 5 && location.getZ() >= getMaxPoint().getZ() - 5;
-        }
-      }
-    }
-    return false;
-  }
-
   public Location getTeleportLocation() {
     try {
-      Location tploc = this.getCenter();
+      Location tploc = cuboid.getCenter();
       while (tploc.getBlock().getType() != Material.AIR || tploc.add(0, 1, 0).getBlock().getType() != Material.AIR) tploc = tploc.add(0, 1, 0);
       boolean enclosed = false;
       int counter = 0;