+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## BuildBattle 3 changelog
 
+### 3.5.2 Release (20.08.2018)
+* Fixed NullPointerException for users who where no longer online
+
 ### 3.5.1 Release (17.08.2018)
 * Fixed NumberFormatException for language.yml migrator - this error is very rare to occur in normal environment but it was reported so I fixed it
 * Fixed NullPointerException for plot adding via selector wand due to lack of code stop when message with not full selection was sent